Add files via upload
Basic application - takes a string input from a html form and translates it to morse code flashed on a LED connected to the rPi.

Store files in the apache webserver document root (/var/www/html by default). Index.php holds the gui (html form), sends a string input in the "Text to encode in morse" textbox to the LED.php script via a jquery AJAX request. The script for the AJAX request is stored in ajaxrq.js. The Morsecode class used for translating plain text strings to a morse code string is stored in morse.php.
The LED is connected to gpio pin 24 (wiring pi pin 5) over a 220Ohm resistor.